src/main/java/app/bladenight/common/procession/Procession.java
package app.bladenight.common.procession;

import app.bladenight.common.events.Event;
import app.bladenight.common.network.messages.EventMessage;
import app.bladenight.common.procession.ProcessionParticipantsListener.ParticipantData;
import app.bladenight.common.procession.Statistics.Segment;
import app.bladenight.common.procession.tasks.ComputeSchedulerClient;
import app.bladenight.common.procession.tasks.ParticipantCollectorClient;
import app.bladenight.common.routes.Route;
import app.bladenight.common.routes.Route.ProjectedLocation;
import app.bladenight.common.time.Clock;
import app.bladenight.common.time.SystemClock;
import org.apache.logging.log4j.LogManager;
import org.apache.logging.log4j.Logger;

import java.util.ArrayList;
import java.util.List;
import java.util.Map;
import java.util.concurrent.ConcurrentHashMap;

public class Procession implements ComputeSchedulerClient, ParticipantCollectorClient {
    public Procession(Clock clock) {
        this.clock = clock;
        init();
    }

    public Procession() {
        init();
    }

    private void init() {
        trackedParticipants = new ConcurrentHashMap<String, Participant>();
        trackedHeads = new ConcurrentHashMap<String, Participant>();
        trackedTails = new ConcurrentHashMap<String, Participant>();
        headMovingPoint = new MovingPoint();
        tailMovingPoint = new MovingPoint();
        route = new Route();
        route.setName("<undefined default route>");

        initComputers();
    }

    private void initComputers() {
        // TODO move to the application configuration
        int nSegments = 200;
        headAndTailComputer = new HeadAndTailComputer(nSegments);
        travelTimeComputer = new TravelTimeComputer(nSegments);
        travelTimeComputer.setClock(clock);
        double routeLength = 0.0;
        if (route != null)
            routeLength = route.getLength();
        if (routeLength <= 0.0) {
            routeLength = 9999.0;
        }
        headAndTailComputer.setRouteLength(routeLength);
        travelTimeComputer.setRouteLength(routeLength);
    }

    public Route getRoute() {
        return route;
    }

    public void setRoute(Route route) {
        if (route == null) {
            getLog().error("setRoute: refusing to set a null route");
            return;
        }
        this.route = route;
        initComputers();
    }

    public EventMessage.EventStatus getEventStatus() {
        return this.eventStatus;
    }

    public void setEventStatus(EventMessage.EventStatus status) {
        this.eventStatus = status;
    }


    // Returned list is read-only!
    public List<Participant> getParticipants() {
        return new ArrayList<Participant>(trackedParticipants.values());
    }

    public void removeParticipant(String deviceId) {
        getLog().debug("Removing participant " + deviceId);
        trackedParticipants.remove(deviceId);
        headAndTailComputer.removeParticipant(deviceId);
        travelTimeComputer.removeParticipant(deviceId);
        trackedHeads.remove(deviceId);
        trackedTails.remove(deviceId);
    }

    public int getParticipantCount() {
        return trackedParticipants.size();
    }

    public int getParticipantsOnRoute() {
        int count = 0;
        for (Participant p : trackedParticipants.values()) {
            if (p.isOnRoute()) {
                count++;
            }
        }
        return count;
    }

    public synchronized Participant updateParticipant(ParticipantInput participantInput) {
        getLog().trace("updateParticipant: " + participantInput);

        String participantId = participantInput.getParticipantId();

        Participant participant;

        // TODO redesign the situation where a participant is not participating (sic)
        if (participantInput.isParticipating()) {
            participant = trackedParticipants.get(participantId);
            if (participant == null) {
                participant = getOrCreateParticipant(participantId);
            }
        } else {
            // we create a temporary participant and we don't add it to the list of tracked participants
            participant = newParticipant(participantId);
        }

        List<ProjectedLocation> potentialLocations = route.projectPosition(participantInput.getLatitude(), participantInput.getLongitude());

        ParticipantUpdater updater = new ParticipantUpdater.Builder().
                setProcessionEnds(getTailPosition(), getHeadPosition()).
                setParticipantInput(participantInput).
                setParticipant(participant).
                setPotentialLocations(potentialLocations).
                setRouteLength(route.getLength()).
                setClock(clock).
                build();

        updater.updateParticipant();

        if (participant.isOnRoute() && participantInput.isParticipating()) {
            ParticipantData participantData = new ParticipantData(participant.getLinearPosition(), participant.getLinearSpeed(), participantInput.getAccuracy());
            headAndTailComputer.updateParticipant(participantId, participantData);
            travelTimeComputer.updateParticipant(participantId, participantData);
        } else {
            headAndTailComputer.removeParticipant(participantId);
            travelTimeComputer.removeParticipant(participantId);
        }

        return participant;
    }

    // Result is read-only
    public Participant getParticipant(String id) {
        return trackedParticipants.get(id);
    }

    private Participant getOrCreateParticipant(String id) {
        Participant p = trackedParticipants.get(id);
        if (p == null) {
            p = newParticipant(id);
            trackedParticipants.put(id, p);
        }
        return p;
    }

    private Participant newParticipant(String id) {
        Participant p = new Participant();
        p.setDeviceId(id);
        return p;
    }

    public boolean isParticipantOnRoute(String deviceId) {
        Participant participant = getParticipant(deviceId);
        if (participant == null)
            return false;
        return participant.isOnRoute();
    }


    public MovingPoint getHead() {
        computeIfTooOld();
        return headMovingPoint;
    }

    public MovingPoint getTail() {
        computeIfTooOld();
        return tailMovingPoint;
    }

    public double getLength() {
        computeIfTooOld();
        return headMovingPoint.getLinearPosition() - tailMovingPoint.getLinearPosition();
    }

    public double getHeadPosition() {
        return getHead().getLinearPosition();
    }

    public double getTailPosition() {
        return getTail().getLinearPosition();
    }

    @Override
    public synchronized void compute() {
        getLog().trace("compute");

        lastComputeTime = clock.currentTimeMillis();

        if (route == null) {
            getLog().error("compute: no route available");
            return;
        }
        double routeLength = route.getLength();
        if (routeLength == 0) {
            getLog().warn("compute: route has zero length -" + route.getName());
            return;
        }

        long startTime = System.currentTimeMillis();

        List<Participant> participantList = new ArrayList<Participant>(trackedParticipants.values());
        getLog().trace("compute: " + participantList.size() + " participants are registered");

        if (!headAndTailComputer.compute()) {
            getLog().trace("compute: could not find the procession position");
            headMovingPoint = new MovingPoint();
            tailMovingPoint = new MovingPoint();
            return;
        }

        MovingPoint newHeadMovingPoint = new MovingPoint();
        MovingPoint newTailMovingPoint = new MovingPoint();

        newHeadMovingPoint.update(0, 0, headAndTailComputer.getHeadPosition());
        newTailMovingPoint.update(0, 0, headAndTailComputer.getTailPosition());

        completeEndMovingPoint(newHeadMovingPoint, headMovingPoint);
        completeEndMovingPoint(newTailMovingPoint, tailMovingPoint);

        headMovingPoint = newHeadMovingPoint;
        tailMovingPoint = newTailMovingPoint;

        getLog().trace("headMovingPoint=" + headMovingPoint);
        getLog().trace("tailMovingPoint=" + tailMovingPoint);

        travelTimeComputer.computeTravelTimeForAllSegments(0.75);

        computeStatistics();

        long endTime = System.currentTimeMillis();

        getLog().trace("compute: compute time: " + (endTime - startTime) + "ms");
    }

    protected void completeEndMovingPoint(MovingPoint newMp, MovingPoint lastMp) {
        newMp.isInProcession(true);
        newMp.isOnRoute(true);

        // Compute the speed of the head or the tail, but don't allow it to jump too much
        if (lastMp != null && lastMp.isOnRoute()) {
            double oldPos = lastMp.getLinearPosition();
            double newPos = newMp.getLinearPosition();
            newMp.setLinearPosition(updateSmoothingFactor * oldPos + (1 - updateSmoothingFactor) * newPos);

            double deltaT = (newMp.getTimestamp() - lastMp.getTimestamp()) / (3600.0 * 1000.0);
            double newSpeed = (newPos - oldPos) / (1000.0 * deltaT);
            newMp.setLinearSpeed(updateSmoothingFactor * lastMp.getLinearSpeed() + (1 - updateSmoothingFactor) * newSpeed);
        }
        Route.LatLong latLong = route.convertLinearPositionToLatLong(newMp.getLinearPosition());
        newMp.setLatLong(latLong.lat, latLong.lon);
    }

    private void computeIfTooOld() {
        if (isMaxComputeAgeEnabled() && isMaxComputeAgeReached()) {
            getLog().debug("Triggering compute automatically");
            compute();
        }
    }

    private boolean isMaxComputeAgeEnabled() {
        return maxComputeAge >= 0;
    }

    private boolean isMaxComputeAgeReached() {
        return lastComputeTime < 0 || maxComputeAge == 0 || System.currentTimeMillis() - lastComputeTime > lastComputeTime;
    }

    public boolean isValid() {
        return getHead().isOnRoute() &&
                getTail().isOnRoute() &&
                getHead().getLinearPosition() >= getTail().getLinearPosition();
    }

    public double getUpdateSmoothingFactor() {
        return updateSmoothingFactor;
    }

    public double evaluateTravelTimeBetween(double position1, double position2) {
        computeIfTooOld();
        double result = travelTimeComputer.evaluateTravelTimeBetween(position1, position2);
        getLog().trace("evaluateTravelTimeBetween(" + position1 + "," + position2 + ")");
        return result;
    }

    public long getMaxComputeAge() {
        return maxComputeAge;
    }

    /*** If set, trigger a compute if the last compute is older than specified.
     * Disabled by default.
     * Set to 0 for systematic updates
     * Set to -1 to disable again.
     */
    public void setMaxComputeAge(long maxComputeAge) {
        this.maxComputeAge = maxComputeAge;
    }


    /***
     * Smoothen jumps of the head and the tail
     * 0.0  : no smoothing
     * 0.99 : unreasonable smoothing
     */
    public void setUpdateSmoothingFactor(double updateSmoothingFactor) {
        this.updateSmoothingFactor = updateSmoothingFactor;
    }

    public double getProcessionGreediness() {
        return headAndTailComputer.getProcessionGreediness();
    }

    public void setProcessionGreediness(double processionGreediness) {
        headAndTailComputer.setProcessionGreediness(processionGreediness);
    }

    public Statistics getStatistics() {
        return statistics;
    }

    public void computeStatistics() {
        int nSegments = travelTimeComputer.getNumberOfSegments();
        SegmentedLinearRoute segmentedLinearRoute = new SegmentedLinearRoute(nSegments, route.getLength());
        double segmentLength = segmentedLinearRoute.getSegmentLength();
        double sumSpeed = 0;
        int nSegmentsWithSpeed = 0;
        Statistics statistics = new Statistics();
        statistics.segments = new Segment[nSegments];
        for (int segment = 0; segment < nSegments; segment++) {
            statistics.segments[segment] = new Segment();
            //          System.out.println("segment="+segment);
            //          System.out.println("segmentLength="+segmentLength);
            //          System.out.println("time="+travelTimeComputer.getTravelTimeForSegment(segment));
            double speed = 3600.0 * (segmentLength / travelTimeComputer.getTravelTimeForSegment(segment));
            statistics.segments[segment].speed = speed;
            if (!Double.isNaN(speed)) {
                sumSpeed += speed;
                nSegmentsWithSpeed++;
            }
        }
        if (nSegmentsWithSpeed > 0)
            statistics.averageSpeed = (sumSpeed / nSegmentsWithSpeed);
        for (Participant p : trackedParticipants.values()) {
            int segment = segmentedLinearRoute.getSegmentForLinearPosition(p.getLinearPosition());
            statistics.segments[segment].nParticipants++;
        }
        this.statistics = statistics;
    }

    private EventMessage.EventStatus eventStatus;
    private Route route;
    private MovingPoint headMovingPoint;
    private MovingPoint tailMovingPoint;
    private HeadAndTailComputer headAndTailComputer;
    private TravelTimeComputer travelTimeComputer;
    private Statistics statistics;
    private long maxComputeAge = -1;
    private long lastComputeTime = -1;
    private Clock clock = new SystemClock();

    protected double updateSmoothingFactor = 0.0;


    private Map<String, Participant> trackedParticipants;


    public Map<String, Participant> trackedHeads;
    public Map<String, Participant> trackedTails;

    private static Logger log;

    public static void setLog(Logger log) {
        Procession.log = log;
    }

    protected static Logger getLog() {
        if (log == null)
            log = LogManager.getLogger(Procession.class.getName());
        return log;
    }

}
